// generated/pagopa-proxy/PaymentActivationsPostRequest.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import { RptId } from "./RptId";
import { ImportoEuroCents } from "./ImportoEuroCents";
import { CodiceContestoPagamento } from "./CodiceContestoPagamento";
import * as t from "io-ts";

/**
 * Define the request received from CD App to require a payment activation
 */

// required attributes
const PaymentActivationsPostRequestR = t.interface({
  rptId: RptId,

  importoSingoloVersamento: ImportoEuroCents,

  codiceContestoPagamento: CodiceContestoPagamento
});

// optional attributes
const PaymentActivationsPostRequestO = t.partial({});

export const PaymentActivationsPostRequest = t.intersection(
  [PaymentActivationsPostRequestR, PaymentActivationsPostRequestO],
  "PaymentActivationsPostRequest"
);

export type PaymentActivationsPostRequest = t.TypeOf<
  typeof PaymentActivationsPostRequest
>;